Fix operator name view reappearing even when user has disabled it
With show_operator_name_in_statusbar_bool enabled the user gets a choice
whether to show or not show the operator name in the status bar.

When the user disables the option in the Settings app, the operator_name
view gets hidden but the CollapsedStatusBarFragment also operates on the
same view so when you drag down the status bar and put it back up then
the operator name reappears even when it's disabled.

Fix this by hiding/unhiding operator_name_frame so the two components
aren't conflicting and not unhiding parts that shouldn't be.

Bug: TBD
Test: enable show_operator_name_in_statusbar_bool, disable "Network
      name" toggle in Settings app, drag QS down, drag up
Change-Id: Ic427f7a04c425bc4c48c89c81a6844e4ef9fdde1
Signed-off-by: minaripenguin <minaripenguin@users.noreply.github.com>